Add open access symbols (#16)
New package for open access symbols in the margin based on a semi-
automatic heuristic.
Symbols drawn with TikZ are available in a support package.
With LuaLaTeX the Unpaywall API can be queried by DOI to obtain a
truly automatic open access detection.

See https://tex.stackexchange.com/q/459449/ and #16