Move from \extblx@DeclareBTEOption to \DeclareBiblatexOption
We implement a fake version of \DeclareBiblatexOption for older
biblatex versions so we can use it directly.
That emulation emits a noisy warning and \DeclareBiblatexOption is
used quite often, so there are going to be quite a lot of warnings,
but I think that is OK.

\DeclareBiblatexOption should be in 3.13,
see also https://github.com/plk/biblatex/pull/866.